% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortem-package.R
\docType{package}
\name{mortem-package}
\alias{mortem}
\alias{mortem-package}
\title{Post-Mortem Debugging}
\description{
Use `mortem` to automatically print a C stack trace when R crashes via
a segfault.
}
